added confirm delete message

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../models/message';
 import { PaginatedResult, Pagination } from '../models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,8 @@ export class MessagesComponent implements OnInit {
   pageNumber = 1;
   isLoading = false;
 
-  constructor(private messageService: MessageService) {}
+
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) {}
 
   ngOnInit(): void {
     this.loadMessages();
@@ -39,8 +41,13 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages?.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('Confirm Delete', 'This cannot be undone').subscribe((result: boolean) => {
+      if(result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages?.splice(this.messages.findIndex(m => m.id === id), 1);
+        });
+      }
     });
+
   }
 }
